news-and-sentiment-agent/requirements.txt
httpx
python-dotenv
uvicorn[standard]
fastapi[standard]
openai
requests
